backend/requirements.txt
Flask==2.3.2
flask-cors==4.0.0
mysql-connector-python==9.1.0
cryptography
PyJWT==2.9.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
